refactor(agent): group the TypeScript runner into role folders + rewrite README
The TS runner's src/ had grown one work package at a time into a flat folder of ten
files with no signal of role. Group them and rewrite the stale README (it still called
this a 'Pi wrapper' and pointed at the moved agent.py):

  src/cli.ts, server.ts, protocol.ts   entrypoints + the wire contract
  src/engines/{pi,rivet}.ts            the two engines (was runPi.ts / runRivet.ts)
  src/tracing/otel.ts                  the tracers (was agenta-otel.ts)
  src/tools/{client,mcp-bridge,mcp-server}.ts   tool delivery (was toolClient/toolBridge*)
  src/extensions/agenta.ts             the Pi extension (was piExtension.ts)

No behavior change. Updated the fragile __dirname-relative paths in engines/rivet.ts
(PKG_ROOT) and tools/mcp-bridge.ts (the tsx bin + server path) for the new depth, and the
build-extension entry. Verified live: rivet+pi+local through the restarted sidecar answers
'Athens' with usage; tsc --strict and the extension build pass.
